Big merge from barsa2000/excludarr (#17)
* Fixed justwatch lib in preparation of sonarr rewrite

I rewrote the queries such that all the offers of each episode i each
season gets retrieved in a single query. I also implemented some basic
filters for the offers query.

* Fixed sonarr

now sonarr works with the new implementation of the justwatch api

some fixes are still needed, but it's a start

* Updated requirements

httpx was needed for the new justwatch api

* Fixed sonarr episodes api

We were using deprecated APIs related to episodes, I updated them

* Fixed TMDB in sonarr

* Fixed justwatch lib in preparation of sonarr rewrite

I rewrote the queries such that all the offers of each episode i each
season gets retrieved in a single query. I also implemented some basic
filters for the offers query.

* Fixed sonarr

now sonarr works with the new implementation of the justwatch api

some fixes are still needed, but it's a start

* Updated requirements

httpx was needed for the new justwatch api

* Fixed sonarr episodes api

We were using deprecated APIs related to episodes, I updated them

* Fixed TMDB in sonarr

* Converted all justwatch requests to httpx

Requests got much faster with this simple conversion

* Typing and Context added

- added better Typing
- moved global variables to a `typer.Context`
- small cleanups
- in every command the parameter `providers`'s type was
  `Option[List[str]]` but if we add the default value of `[]` it should
  no longer be `Option[...]`, so it was fixed

* Format

* Fixed a type

* more types

* Flake8

* Added flake8 and black mypy to pre-commit-config

* More Flake8 and formatting

* Converted to poetry

* Fixed Dockerfile to work with poetry

* Fixes related to the various merges

* Pre-commit and project fixes

* Removed unused fn

* fixed Dockerfile

* Added license checker

* Added log file for better debugging

* Added tag based skip-list functionality

* updated readme and examples

* Added basic python program to populate *arr

* Added `arr-setup` program to populate radarr and sonarr based on a given
config.

* Added `arr_setup` as dev dependency

* added support for season/ep control for each show

now we can control each episode that gets added to the testing
environment from the config, we even have a wildcard :)

* Added example for arr-setup

* Added config flag to setup command

* dockerized everything and setup dev env

config_arr_setup was added as a dependency to extract data from radarr
and sonarr config and putting it in arr-setup config, so it is fully
automated, we just have to copy the config we want to use for arr-setup
to its volume and just `docker compose up` :)

* update docker workflow and version

* Format docker workflow

* add workflow dispatch to docker

* Fix

* Updated README and added verification to docker workflow

* Fix for arm build

* Updated README

* Fix: fixed bug in dockerfile

There was a bug in the entrypoint where the args passed to it were not
passed to excludarr

* Added developer info to run project

* Fixed LICENSE

* FIX: changed all refs from old repo to excludarr/excludarr

* FIX: removed all refs to simplejustwatchapi

During the merge I didn't notice it was added back a single reference to
that lib.

* CHORE: updated version in pyproject

* CHORE: updated lockfile

* FIX: fixed refs to old dockerhub repo

* BUILD: pinned image and packages versions

* Fix: format, black, flake8

* FIX: removed unnecessary print